requirements.txt
# IT Help Desk - All Python Dependencies
azure-ai-projects>=1.0.0
azure-ai-agents>=1.0.0b1
azure-identity>=1.19.0
azure-cosmos>=4.15.0
python-dotenv>=1.0.1
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
